Skipping another DROP TABLE (missed out from previous commit).

+++ core/src/test/java/io/questdb/cutlass/pgwire/PGJobContextTest.java
@@ -6209,7 +6209,9 @@ create table tab as (
                     assertEquals(expectedTs, tsBack);
 
                     // cleanup
-                    conn.prepareStatement("drop table ts").execute();
+                    if (!isDisabledForWalRun()) {
+                        conn.prepareStatement("drop table ts").execute();
+                    }
                 }
             }
         });
